pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eduardo Beckhauser", email = ""},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

